[CEN-1442] Added checkstyle to Maven lifecycle - Partially fixed issues
