+++ pom.xml
@@ -10,9 +10,9 @@
     <!-- lookup parent from repository -->
   </parent>
   <groupId>it.gov.pagopa.rtd.ms</groupId>
-  <artifactId>rtd-ms-decrypter</artifactId>
+  <artifactId>rtd-ms-decrypterImpl</artifactId>
   <version>1.0.1</version>
-  <name>rtd-ms-decrypter</name>
+  <name>rtd-ms-decrypterImpl</name>
   <description>Microservice capable of decrypting PGP files</description>
   <properties>
     <java.version>11</java.version>
@@ -88,6 +88,34 @@
       <plugin>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-maven-plugin</artifactId>
+      </plugin>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-checkstyle-plugin</artifactId>
+        <version>3.1.2</version>
+        <dependencies>
+          <dependency>
+            <groupId>com.puppycrawl.tools</groupId>
+            <artifactId>checkstyle</artifactId>
+            <version>10.0</version>
+          </dependency>
+        </dependencies>
+        <executions>
+          <execution>
+            <id>validate</id>
+            <phase>validate</phase>
+            <goals>
+              <goal>check</goal>
+            </goals>
+          </execution>
+        </executions>
+        <configuration>
+          <configLocation>google_checks.xml</configLocation>
+          <encoding>UTF-8</encoding>
+          <consoleOutput>true</consoleOutput>
+          <failsOnError>true</failsOnError>
+          <linkXRef>false</linkXRef>
+        </configuration>
       </plugin>
     </plugins>
   </build>
