+++ src/main/java/it/gov/pagopa/rtd/ms/rtdmsdecrypter/configuration/ThreadSafeHttpClient.java
@@ -1,5 +1,9 @@
 package it.gov.pagopa.rtd.ms.rtdmsdecrypter.configuration;
 
+import java.security.KeyManagementException;
+import java.security.KeyStoreException;
+import java.security.NoSuchAlgorithmException;
+import javax.net.ssl.SSLContext;
 import org.apache.http.config.Registry;
 import org.apache.http.config.RegistryBuilder;
 import org.apache.http.conn.socket.ConnectionSocketFactory;
@@ -14,26 +18,24 @@ import org.apache.http.ssl.SSLContexts;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import javax.net.ssl.SSLContext;
-import java.security.KeyManagementException;
-import java.security.KeyStoreException;
-import java.security.NoSuchAlgorithmException;
-
 @Configuration
 public class ThreadSafeHttpClient {
 
   @Bean
-  CloseableHttpClient myHttpClient() throws NoSuchAlgorithmException, KeyStoreException, KeyManagementException {
+  CloseableHttpClient myHttpClient()
+      throws NoSuchAlgorithmException, KeyStoreException, KeyManagementException {
     SSLContext sslContext = SSLContexts.custom()
-            .loadTrustMaterial(TrustSelfSignedStrategy.INSTANCE)
-            .build();
+        .loadTrustMaterial(TrustSelfSignedStrategy.INSTANCE)
+        .build();
 
-    Registry<ConnectionSocketFactory> registry = RegistryBuilder.<ConnectionSocketFactory> create()
-            .register("http", PlainConnectionSocketFactory.INSTANCE)
-            .register("https", new SSLConnectionSocketFactory(sslContext, NoopHostnameVerifier.INSTANCE))
-            .build();
+    Registry<ConnectionSocketFactory> registry = RegistryBuilder.<ConnectionSocketFactory>create()
+        .register("http", PlainConnectionSocketFactory.INSTANCE)
+        .register("https",
+            new SSLConnectionSocketFactory(sslContext, NoopHostnameVerifier.INSTANCE))
+        .build();
 
-    PoolingHttpClientConnectionManager connectionManager = new PoolingHttpClientConnectionManager(registry);
+    PoolingHttpClientConnectionManager connectionManager = new PoolingHttpClientConnectionManager(
+        registry);
 
     return HttpClients.custom().setConnectionManager(connectionManager).build();
   }
