--- src/main/java/it/gov/pagopa/rtd/ms/rtdmsdecrypter/service/IBlobRestConnector.java
@@ -1,13 +1,0 @@
-package it.gov.pagopa.rtd.ms.rtdmsdecrypter.service;
-
-import it.gov.pagopa.rtd.ms.rtdmsdecrypter.model.BlobApplicationAware;
-
-public interface IBlobRestConnector {
-
-  // Downloads a blob and returns the name of the downloaded file
-  BlobApplicationAware get(BlobApplicationAware blobUri);
-
-  // Uploads a blob and returns the name of the uploaded file
-  BlobApplicationAware put(BlobApplicationAware fileName);
-
-}
