--- src/main/java/it/gov/pagopa/rtd/ms/rtdmsdecrypter/service/IDecrypter.java
@@ -1,8 +1,0 @@
-package it.gov.pagopa.rtd.ms.rtdmsdecrypter.service;
-
-import it.gov.pagopa.rtd.ms.rtdmsdecrypter.model.BlobApplicationAware;
-
-public interface IDecrypter {
-  BlobApplicationAware decrypt(BlobApplicationAware blob);
-
-}
